# docs/swagger.yaml
basePath: /
definitions:
  INF368_internal_models.ErrorDetail:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  INF368_internal_models.ErrorResponse:
    properties:
      errors:
        items:
          $ref: '#/definitions/INF368_internal_models.ErrorDetail'
        type: array
    type: object
  INF368_internal_models.Input:
    properties:
      title:
        type: string
    required:
    - title
    type: object
  INF368_internal_models.SuccessResponse:
    properties:
      result: {}
    type: object
externalDocs:
  description: OpenAPI
  url: https://swagger.io/resources/open-api/
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server for managing books.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Book API
  version: "1.0"
paths:
  /books:
    get:
      description: This endpoint retrieves all books in the system.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/INF368_internal_models.SuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/INF368_internal_models.ErrorResponse'
      summary: Get all books
      tags:
      - books
    post:
      consumes:
      - application/json
      description: This endpoint creates a new book with a given title.
      parameters:
      - description: Book title
        in: body
        name: title
        required: true
        schema:
          $ref: '#/definitions/INF368_internal_models.Input'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/INF368_internal_models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/INF368_internal_models.ErrorResponse'
      summary: Create a new book
      tags:
      - books
  /books/{id}:
    delete:
      description: This endpoint deletes a book by its ID.
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/INF368_internal_models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/INF368_internal_models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/INF368_internal_models.ErrorResponse'
      summary: Delete a book
      tags:
      - books
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"